+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 * `PayPalNativePayments`:
   *  Bump `PayPal Native Checkout` to `0.8.8` and add `return_url`
+* Send analytic events for `PayPalNativePayments`, `PayPalWebPayments`, and `CardPayments` flows
 
 ## 0.0.7 (2023-01-25)
 * Rename `PayPalDataCollector` to `FraudProtection`